make it better

+++ app/Http/Controllers/PostController.php
@@ -5,6 +5,8 @@ use Illuminate\Support\Facades\Auth;
 use Illuminate\Http\Request;
 use App\Post;
 use Illuminate\Support\Facades\DB;
+use Illuminate\Support\Facades\Storage;
+
 class PostController extends Controller
 {
     
@@ -35,13 +37,14 @@ class PostController extends Controller
     {
        
 
-        $post = Post::orderBy('created_at','DESC')->paginate(10);
+        $post = Post::orderBy('created_at','desc')->paginate(10);
 
         return view('posts.index', compact('post'));
 
 
 
     }
+    
 
     /**
      * Show the form for creating a new resource.
@@ -66,8 +69,27 @@ class PostController extends Controller
             'name'=>'required',
             'body'=>'required',
             'email'=>'required',
+            'post_image'=>'image|nullable|max:1024',
+
               ]);
 
+
+       if($request->hasFile('post_image')){  
+
+    $filenameWithExtention = $request->file('post_image')->getClientOriginalName();
+    $fileName = pathinfo($filenameWithExtention,PATHINFO_FILENAME);
+    $extension = $request->file('post_image')->getClientOriginalExtension();
+    $fileNameStore = $fileName .'_'.time().'.'.$extension;
+
+    $path = $request->file('post_image')->move(base_path() . '/public/images/', $fileNameStore);
+  
+ 
+            
+        }else{
+                $fileNameStore = 'noImage.jpg';
+              } 
+
+        
               $user = Auth::user();
               $post = new Post();
               $post->name = $request->input('name');
@@ -77,6 +99,7 @@ class PostController extends Controller
               $post->tag = $request->input('tag'); 
               $post->more = $request->input('more'); 
               $post->user_id = $user->id;
+              $post->post_image = $fileNameStore;
 
               $post->save();
 
@@ -118,23 +141,39 @@ class PostController extends Controller
      * @param  int  $id
      * @return \Illuminate\Http\Response
      */
+    
     public function update(Request $request, $id)
     {
-    
-        $request->validate([
+        $this->validate($request ,[
             'name'=>'required',
-            'body'=>'required',
-            'email'=>'required'
-              ]);
+            'email'=>'required',
+            'body'=>'required'
+           ]);
 
-              $post =  Post::find($id);
-              $post->name = $request->input('name');
-              $post->body = $request->input('body');    
-              $post->email = $request->input('email');    
-              $post->save();
-
-              return redirect('posts')->with('success' , 'تم تعديل البلاغ بنجاح ');
+           if($request->hasFile('post_image')){
+            $filenameWithExtention = $request->file('post_image')->getClientOriginalName();
+            $fileName = pathinfo($filenameWithExtention,PATHINFO_FILENAME);
+            $extension = $request->file('post_image')->getClientOriginalExtension();
+            $fileNameStore = $fileName .'_'.time().'.'.$extension;
+          
+            $path = $request->file('post_image')->move(base_path() . '/public/images/', $fileNameStore);
+            // $path = $request->file('post_image')->storeAs('public/post_image',$fileNameStore);
+        } 
+        $post =   Post::find($id);
+        $post->name = $request->input('name');
+        $post->email = $request->input('email');  
+        $post->body  = $request->input('body');
+        $post->tag = $request->input('tag');
+        $post->body      = $request->input('body');
+        $post->more      = $request->input('more');
+        if($request->hasFile('post_image')){
+            $post->post_image      =   $fileNameStore;
+        }
+        $post->save();
+        
+      return redirect('/home')->with('success', 'Done successfully');
     }
+
 
     /**
      * Remove the specified resource from storage.
@@ -144,9 +183,18 @@ class PostController extends Controller
      */
     public function destroy($id)
     {
-        $post = Post::find($id);
-        $post ->delete();
-        return redirect('/posts')->with('success' , 'تم حذف البلاغ بجاح');
+        $post =   Post::find($id);
+        if(auth()->user()->id !== $post->user_id){
+            return redirect('/posts')->with('error','Unauthorized');
+        }
+        if($post->post_image != 'noImage.jpg'){
+            Storage::delete('public/images/'.$post->post_image);
+        
+        }
 
+
+                $post->delete() ;   
+            
+                return redirect('/posts')->with('success', 'Done successfully');
     }
 }
